backend/pyproject.toml
[tool.poetry]
name = "jurba-media-django-backend"
version = "0.1.0"
description = ""
authors = ["Demaga <demaga@i.ua>"]
readme = "README.md"
# packages = [{include = "jurba_media_django_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.21.0"
django-storages = "^1.13.1"
martor = "^1.6.14"
Pillow = "^9.2.0"
boto3 = "^1.24.89"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-colorfield = "^0.7.2"
Faker = "^15.1.1"
python-slugify = "^6.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
